+++ sys/include/ape/u.h
@@ -9,10 +9,10 @@ typedef	unsigned short	ushort;
 typedef	unsigned char	uchar;
 typedef unsigned long	ulong;
 typedef unsigned int	uint;
-typedef   signed char	schar;
+typedef signed char	schar;
 typedef	long long	vlong;
 typedef	unsigned long long uvlong;
-typedef	ushort		Rune;
+typedef	uint		Rune;
 typedef 	union FPdbleword FPdbleword;
 typedef	char*	p9va_list;
